+++ tests/bootstrap.php
@@ -11,5 +11,6 @@ require_once __DIR__ . '/../vendor/autoload.php';
 define('TESTS_ROOT', __DIR__);
 
 require_once TESTS_ROOT . '/classes/app/controller/Test6547586Ctl.php';
+require_once TESTS_ROOT . '/classes/app/controller/Main6547586Ctl.php';
 require_once TESTS_ROOT . '/classes/admin/controller/Test6547586Ctl.php';
 require_once TESTS_ROOT . '/classes/dto/TestDTO.php';