+++ tests/classes/app/controller/Main6547586Ctl.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Created by IntelliJ IDEA.
+ * User: alex
+ * Date: 02.08.17
+ * Time: 22:05
+ */
+
+namespace app\controller;
+
+use microapi\Controller;
+
+class Main6547586Ctl extends Controller {
+
+    /**
+     * @return array
+     * @methods(get)
+     */
+    public function actionIndex() {
+        return ['prompt' => 'default action from default controller'];
+    }
+}