+++ tests/data/mods_endpoints_get.php
@@ -6,6 +6,13 @@
  */ 
 
 return [
+  'app\controller\Main6547586Ctl' => [
+    'index' => [
+      'methodName' => 'actionIndex',
+      'paramsMeta' => [
+      ],
+    ],
+  ],
   'app\controller\Test6547586Ctl' => [
     'index' => [
       'methodName' => 'actionIndex',